Started rethinking the structure
